+++ libass/ass_font.h
@@ -40,7 +40,7 @@ typedef struct ass_font_s {
 	int n_faces;
 	double scale_x, scale_y; // current transform
 	FT_Vector v; // current shift
-	int size;
+	double size;
 #ifdef HAVE_FONTCONFIG
 	FcCharSet* charset;
 #endif
@@ -48,7 +48,7 @@ typedef struct ass_font_s {
 
 ass_font_t* ass_font_new(ass_library_t* library, FT_Library ftlibrary, void* fc_priv, ass_font_desc_t* desc);
 void ass_font_set_transform(ass_font_t* font, double scale_x, double scale_y, FT_Vector* v);
-void ass_font_set_size(ass_font_t* font, int size);
+void ass_font_set_size(ass_font_t* font, double size);
 void ass_font_get_asc_desc(ass_font_t* font, uint32_t ch, int* asc, int* desc);
 FT_Glyph ass_font_get_glyph(void* fontconfig_priv, ass_font_t* font, uint32_t ch, ass_hinting_t hinting);
 FT_Vector ass_font_get_kerning(ass_font_t* font, uint32_t c1, uint32_t c2);
